Home page and 3 recipes added
Forgot to commit as I went so this will add the base format of the
home page, as well as the bare bones of feta stuffed chicken,
chocolate mousse, and smoked salmon scrambled eggs.
They will have a title, an image, short desciption, ingredients list
and method.
Each of the 3 pages will also have a link back to home page.